+++ src/Maxar_Portal_SDK/account_service/activations.py
@@ -6,7 +6,7 @@ import Maxar_Portal_SDK.process as process
 class Activations:
 
     def __init__(self, auth):
-        self.base_url = auth.base_url
+        self.base_url = auth.api_base_url
         self.response = None
         self.version = auth.version
         self.auth = auth
@@ -23,7 +23,7 @@ class Activations:
 
         authorization = process.authorization(self.auth)
         url = self.base_url + '/accountservice/api/v1/activations?search={}'.format(search)
-        response = requests.request("GET", url, headers=authorization, verify=False)
+        response = requests.request("GET", url, headers=authorization, verify=self.auth.SSL)
         process_response = process._response_handler(response)
         if len(response.json()["activations"]) < 1:
             raise Exception("No search results for {}. Please try another search term".format(search))
@@ -44,7 +44,7 @@ class Activations:
             url = self.base_url + '/accountservice/api/v1/activations/{}'.format(activationId)
         elif not activationId:
             url = self.base_url + '/accountservice/api/v1/activations'
-        response = requests.request("GET", url, headers=authorization, verify=False)
+        response = requests.request("GET", url, headers=authorization, verify=self.auth.SSL)
         process_response = process._response_handler(response)
         return response.json()
 
@@ -60,7 +60,7 @@ class Activations:
 
         authorization = process.authorization(self.auth)
         url = self.base_url + '/accountservice/api/v1/activations/available?accountId={}'.format(accountId)
-        response = requests.request("GET", url, headers=authorization, verify=False)
+        response = requests.request("GET", url, headers=authorization, verify=self.auth.SSL)
         process_response = process._response_handler(response)
         return response.json()
 
@@ -74,7 +74,7 @@ class Activations:
 
         authorization = process.authorization(self.auth)
         url = self.base_url + '/accountservice/api/v1/activations/roles/{}'.format(activationId)
-        response = requests.request("GET", url, headers=authorization, verify=False)
+        response = requests.request("GET", url, headers=authorization, verify=self.auth.SSL)
         process_response = process._response_handler(response)
         return response.json()
 
@@ -87,7 +87,7 @@ class Activations:
 
         authorization = process.authorization(self.auth)
         url = self.base_url + '/accountservice/api/v1/activations/types'
-        response = requests.request("GET", url, headers=authorization, verify=False)
+        response = requests.request("GET", url, headers=authorization, verify=self.auth.SSL)
         process_response = process._response_handler(response)
         return response.json()
 
@@ -171,7 +171,7 @@ class Activations:
                 payload[item] = kwargs[item]
 
         payload = json.dumps(payload)
-        response = requests.request("POST", url, headers=authorization, data=payload, verify=False)
+        response = requests.request("POST", url, headers=authorization, data=payload, verify=self.auth.SSL)
         process_response = process._response_handler(response)
         return response.json()
 
@@ -234,7 +234,7 @@ class Activations:
                 payload[item] = kwargs[item]
 
         payload = json.dumps(payload)
-        response = requests.request("PUT", url, headers=authorization, data=payload, verify=False)
+        response = requests.request("PUT", url, headers=authorization, data=payload, verify=self.auth.SSL)
         process_response = process._response_handler(response)
         return response.json()
 
@@ -249,6 +249,6 @@ class Activations:
 
         authorization = process.authorization(self.auth)
         url = self.base_url + '/accountservice/api/v1/activations?id={}'.format(activationId)
-        response = requests.request("DELETE", url, headers=authorization, verify=False)
+        response = requests.request("DELETE", url, headers=authorization, verify=self.auth.SSL)
         process_response = process._response_handler(response)
         return "Activation {} successfully deleted".format(activationId)