+++ docs/release_notes.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## Version 1.1.0 (November 16, 2022)
+
+* Updated API endpoints, base url no longer needed in config file.
+* Added CQL filter validation
+
 ## Version 1.0.0 (October 28, 2022)
 
 * Initial release